refactor: remove GameHUD and move stats to BottomDock
